backend/requirements.txt
Django~=6.0
python-dotenv~=1.2.1
pillow~=10.4.0
rest_framework_simplejwt~=0.0.2
djangorestframework-simplejwt
django-cors-headers
channels~=4.3.2
daphne~=4.2.1
django-organizations~=2.5.0